+++ users/models.py
@@ -10,7 +10,7 @@ from phone_field import PhoneField
 access_level = [(1,'basic'),(2,'Intermediate'),(3,'Admin')]
 
 class user_extended(models.Model):
-    access_level = access_level = models.IntegerField(choices=access_level,default=1)
+    access_level = access_level = models.PositiveSmallIntegerField(choices=access_level,default=1)
     User = models.OneToOneField(User,on_delete=models.CASCADE)
     address = models.CharField(max_length=100,default="")
     mobile = PhoneField(blank=True,help_text="Mobile_Number")
